requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
cryptography==41.0.7
datasets==4.1.1
dbus-python==1.3.2
dill==0.4.0
distlib==0.4.0
distro==1.9.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
hf-xet==1.1.10
httplib2==0.20.4
huggingface-hub==0.35.3
idna==3.10
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
multidict==6.6.4
multiprocess==0.70.16
numpy==2.3.3
oauthlib==3.2.2
packaging==25.0
pandas==2.3.3
platformdirs==4.4.0
propcache==0.3.2
pyarrow==21.0.0
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
python-apt==2.7.7+ubuntu5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.16.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
wadllib==1.3.6
xxhash==3.5.0
yarl==1.20.1
